requirements.txt
# Core
Flask>=3.0.0
python-dotenv>=1.0.0
git+https://github.com/openai/whisper.git  
torch>=2.0.0 --extra-index-url https://download.pytorch.org/whl/cu118
torchaudio>=2.0.0

# Audio
sounddevice>=0.4.6; sys_platform != 'win32'
pyaudio>=0.2.13; sys_platform == 'win32'

# OpenAI
openai>=1.0.0
